Shared Tauri-free startup/shutdown in meetily-core::bootstrap
Move the startup/shutdown sequence out of the Tauri shell's lib.rs into a
new meetily-core::bootstrap module so the upcoming GPUI shell can reuse it
verbatim: init_paths (models + templates dirs), prepare_database (wraps
database::setup::prepare_database_on_startup), spawn_background_init
(whisper init, speaker-diarizer pre-warm, summary ModelManager init, model
audit log, missing-model downloads), finish_recording_for_exit (the
mid-recording close/exit flow), and shutdown (DB checkpoint, sidecar
teardown, whisper unload — _exit itself stays shell-side).

Fixes a startup race along the way: the speaker-diarizer pre-warm used to
be spawned before the database was opened (block_on'd later in setup),
logging "DB pool unavailable; cannot load voice profiles" on every normal
launch. The Tauri shell's setup() now runs init_paths() -> prepare_database()
-> manage(AppState) -> spawn_background_init(sink, pool, model_manager_slot)
in that order, so the pre-warm gets the real pool.

The Tauri shell's lib.rs setup/exit handlers and begin_shutdown_stop are now
thin calls into bootstrap; recording_commands::build_context is pub(crate)
so begin_shutdown_stop can build the same tray-refreshing RecordingContext
finish_recording_for_exit needs.
