+++ meetily-core/src/lib.rs
@@ -22,6 +22,7 @@ macro_rules! perf_debug {
 
 pub mod anthropic;
 pub mod audio;
+pub mod bootstrap;
 pub mod calendar;
 pub mod config;
 pub mod database;
